♻️ [refactor] 오버레이모달 합성컴포넌트 작업
