+++ apps/web/src/shared/ui/common/OverlayModal/Text.tsx
@@ -1,0 +1,42 @@
+import {
+  DialogHeader,
+  DialogTitle,
+  DialogDescription,
+} from '@/shared/ui/common/dialog'
+import { cn } from '@/shared/lib/utils'
+import type { ComponentProps } from 'react'
+
+interface TextProps extends ComponentProps<'div'> {
+  heading: string
+  body?: string
+  isImageModal?: boolean
+}
+
+export const Text = ({
+  heading,
+  body,
+  isImageModal,
+  className,
+  ...rest
+}: TextProps) => {
+  return (
+    <DialogHeader>
+      <div
+        className={cn(
+          'flex flex-col space-y-2 text-center items-center',
+          isImageModal && 'mb-8',
+          className,
+        )}
+        {...rest}
+        style={{ color: 'var(--color-usage-text-inverse)' }}
+      >
+        <DialogTitle className="body-lg-bold">{heading}</DialogTitle>
+        {body && (
+          <DialogDescription className="body-sm-medium">
+            {body}
+          </DialogDescription>
+        )}
+      </div>
+    </DialogHeader>
+  )
+}
